requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.13
fastapi-jwt-auth==0.5.0

starlette==0.27.0
sqlalchemy==1.4.49
psycopg2-binary==2.9.10

passlib[bcrypt]==1.7.4
bcrypt==4.1.2

email-validator==1.3.1
python-dotenv==1.0.1
PyJWT==1.7.1
PyYAML==6.0.2

anyio==3.6.2
click==8.1.7
idna==3.7
sniffio==1.3.0
typing-extensions==4.12.0